+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## [Unreleased](https://github.com/Piturnah/gex/compare/v0.3.7...master)
+### Added
+- Support for <kbd>Del</kbd> in minibuffer
 ### Changed
-- Cursor switches to bar when navigating left and right in git command
+- Cursor switches to bar when navigating left and right in minibuffer
 
 ## [0.3.7](https://github.com/Piturnah/gex/compare/v0.3.6...v0.3.7) - 2023-01-25
 ### Added